requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aenum==3.1.15
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1; platform_python_implementation != "PyPy"
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
cryptography==43.0.3
distlib==0.3.9
dj-database-url==2.3.0
django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
identify==2.6.1
idna==3.10
importlib-metadata==8.5.0; python_version < "3.12"
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
netifaces==0.11.0
nodeenv==1.9.1
packaging==24.1
passage-identity==2.5.1
pillow==11.0.0
platformdirs==4.3.6
pre-commit==4.0.1
psycopg2-binary==2.9.10
pycparser==2.22; platform_python_implementation != "PyPy"
pydantic==2.9.2
pydantic-core==2.23.4
pydotplus==2.0.2
pyjwt==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27; sys_platform == "linux"
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.32.0
virtualenv==20.27.1
whitenoise==6.8.2
zipp==3.20.2; python_version < "3.12"
